refactor: create Navbar component
